HopTB.Floquet

HopTB.Floquet.FTBModel — Type

mutable struct FTBModel{T<:Number}
    Ω::Float64
    tm::TBModel{T}
    num_independent_orbits::Int64
    forder::Int64
end

FTBModel is a struct representing a Floquet tight binding model, which is is a tight binding model (whose number of orbits being num_independent_orbits) with a periodic Hamiltonian (frequency being Ω). A Floquet tight binding model can be mapped to a tight binding model with an electric field, stored as tm. tm should in principle have infinite number of orbits and we cut off tm from -forder to forder in the frequency dimension.

The basis function for tm is |n⟩e^{-ilΩt}, where n labels tight binding orbitals of the original model. The basis function above corresponds to the (l+forder)*num_independent_orbits+n row/column in the matrix.

HopTB.Floquet.getFTBModel — Function

function getFTBModel(tm::TBModel{T}, Ω::Float64, forder::Int64;
    keep_original_hoppings::Bool=false) where T<:Number

Construct a FTBModel from a TBModel.

HopTB.Floquet.addlight — Function

A(t) = real(A*exp(iΩt))

⟨0n|H'(t)|Rm⟩ = ⟨0n|H|Rm⟩exp(im(A(t)⋅(R+rm-rn))) where the negative charge of electron has been taken into account.