HopTB.Hall

HopTB.Hall.getahc — Function

getahc(atm::AbstractTBModel, α::Int64, β::Int64, nkmesh::Vector{Int64};
    Ts::Vector{Float64}=[0.0], μs::Vector{Float64}=[0.0])::Matrix{Float64}

Calculate anomalous Hall conductivity $σ^{αβ}$.

Anomalous Hall conductivity is defined by

\[σ^{αβ}=-\frac{e^2}{ħ}\int\frac{d\boldsymbol{k}}{(2pi)^3}f_nΩ_{nn}^{αβ}.\]

The returned matrix $σ^{αβ}[m, n]$ is AHC for temperature Ts[m] and chemical potential μs[n].

The returned AHC is in unit (Ω⋅cm)^-1.

HopTB.Hall.getshc — Function

getshc(tm::TBModel, α::Int64, β::Int64, γ::Int64, nkmesh::Vector{Int64};
    Ts::Vector{Float64}=[0.0], μs::Vector{Float64}=[0.0], ϵ::Float64=0.1)::Matrix{Float64}

Calculate spin Hall conductivity for different temperature (Ts, first dimension) and chemical potential (μs, second dimension).

Spin Hall conductivity is defined as

\[σ_{αβ}^{γ} = eħ\int\frac{d^3 \boldsymbol{k}}{(2π)^3}\sum_n f_n Ω^{γ}_{n,αβ},\]

where the spin Berry curvature is

\[Ω_{n,αβ}^{γ} = -2 \text{Im} [\sum_{m≠n} \frac{⟨n|\hat{j}_α^γ|m⟩⟨m|\hat{v}_β|n⟩}{(ϵ_n-ϵ_m)^2+ϵ^2}]\]

and the spin current operator is

\[\hat{j}_α^γ = \frac{1}{2} \{\hat{v}_a, \hat{s}_c\}.\]

Spin Hall conductivity from this function is in ħ/e (Ω*cm)^-1.

HopTB.Hall.collect_berry_curvature — Function

collect_berry_curvature(atm::AbstractTBModel, α::Int64, β::Int64, kpts::AbstractMatrix{Float64})::Matrix{Float64}

Collect berry curvature.

Standard units is used (eV and Å).

The returned matrix Ω[n, ik] is berry curvature for band n at ik point.