HopTB.Magnetism

HopTB.Magnetism.get_orbital_moment — Function

get_orbital_moment(tm::AbstractTBModel, α::Int64, k::Vector{Float64})
    --> M::Matrix{ComplexF64}

This function returns orbital magnetic moment in the unit of Bohr magneton

\[M_{nm}=i m ϵ^{αβγ} \sum_{\bar{l}} \frac{v_{nl}^β v_{lm}^γ}{ϵ_n-ϵ_l},\]

where either n = m or band n and band m are completely degenerate.

M[n, m] would be set to zero if the above condition is not fulfilled.

Warning

This function has not been tested.

HopTB.Magnetism.get_field_modified_Es — Function

get_field_modified_Es(tm::AbstractTBModel, α::Int64, B::Float64,
    k::Vector{Float64}; double_degenerate=false)
    --> mEs::Vector{Float64}

Calculate modified energy due to orbital and spin Zeeman correction. If double_degenerate is true, the bands are assumed to be doubly degenerate. Otherwise the bands are assumed to be nondegenerate.

Warning

This function has not been tested.