HopTB.Response

HopTB.Response.getresponse — Function

getresponse(getH::Function, getO::Function, getV::Function,
    getn::Function, k::AbstractVector{Float64}; Γ::Float64=0.1, atol::Float64=0.0,
    maxevals::Int64=typemax{Int64})

Calculate response at k.

The response is the linear change of ⟨O⟩ due to V added to the original Hamiltonian. getH, getO and getV all take k as the only argument and output corresponding matrices.

Γ is the broadening factor. If the every degree of the system is coupled to a wide-band heat bath, Γ=i T^2 ν, where T is the hopping constant between the system and the heat bath and ν is the density of states of the heat bath. The heat bath has a occupation distribution getn(ω).

The integration over ω would have an absolute tolerance atol, with maximal evaluation number being maxevals.

getresponse(getH::Function, getO::Function, getV::Function,
    getn::Function, nkmesh::Vector{Int64}; Γ::Float64=0.1, atol::Float64=0.0,
    maxevals::Int64=typemax(Int64), offset::Vector{Float64}=[0.0, 0.0, 0.0],
    k1::Vector{Float64}=[0.0, 0.0, 0.0], k2::Vector{Float64}=[1.0, 1.0, 1.0])

Calculate response.

The response is the linear change of ⟨O⟩ due to V added to the original Hamiltonian. getH, getO and getV all take k as the only argument and output corresponding matrices.

Γ is the broadening factor. If the every degree of the system is coupled to a wide-band heat bath, Γ=i T^2 ν, where T is the hopping constant between the system and the heat bath and ν is the density of states of the heat bath. The heat bath has a occupation distribution getn(ω).

The response function is integrated in a rectangle with corners k1 and k2. The mesh is nkmesh and every k point is offsetted by offset.

The integration over ω would have an absolute tolerance atol, with maximal evaluation number being maxevals for each k point.

For a crystal, since this function does not know the volume of the Brillouin zone, it simply sums over k points and divide the result by the number of k points. The most common situation is calculating the density of ⟨O⟩, in which case the result should be multiplied by the volume of the Brillouin zone and divided by (2π)^3.

HopTB.Response.get_Floquet_response — Function

get_Floquet_response(getH::Function, getO::Function, getV::Function,
    getn::Function, l::Int64, k::Vector{Float64}; Γ::Float64=0.1,
    atol::Float64=0.0, maxevals::Int64=typemax(Int64))

Calculate Floquet response at k.

The response is the linear change of ⟨O⟩ due to V added to the original Hamiltonian. getH, getO and getV all take k as the only argument and output corresponding matrices.

We here assume operator O and V are time independent, such that their size should be number of orbital. However, H should be Floquet Hamiltonian and should have size (2forder+1)norbits.

⟨O⟩ might be time dependent but periodic. l denotes that the output is the coefficient of ⟨O⟩ before e^{-i l Ω t}.

Γ is the broadening factor. If the every degree of the system is coupled to a wide-band heat bath, Γ=i T^2 ν, where T is the hopping constant between the system and the heat bath and ν is the density of states of the heat bath. The heat bath has a occupation distribution getn(ω).

The integration over ω would have an absolute tolerance atol, with maximal evaluation number being maxevals.

get_Floquet_response(getH::Function, getO::Function, getV::Function,
    getn::Function, l::Int64, nkmesh::Vector{Int64}; Γ::Float64=0.1,
    offset::Vector{Float64}=[0.0, 0.0, 0.0], k1::Vector{Float64}=[0.0, 0.0, 0.0],
    k2::Vector{Float64}=[1.0, 1.0, 1.0], atol::Float64=0.0, maxevals::Int64=typemax(Int64))

Calculate Floquet response.

The response is the linear change of ⟨O⟩ due to V added to the original Hamiltonian. getH, getO and getV all take k as the only argument and output corresponding matrices.

We here assume operator O and V are time independent, such that their size should be number of orbital. However, H should be Floquet Hamiltonian and should have size (2forder+1)norbits.

⟨O⟩ might be time dependent but periodic. l denotes that the output is the coefficient of ⟨O⟩ before e^{-i l Ω t}.

Γ is the broadening factor. If the every degree of the system is coupled to a wide-band heat bath, Γ=i T^2 ν, where T is the hopping constant between the system and the heat bath and ν is the density of states of the heat bath. The heat bath has a occupation distribution getn(ω).

The response function is integrated in a rectangle with corners k1 and k2. The mesh is nkmesh and every k point is offsetted by offset.

The integration over ω would have an absolute tolerance atol, with maximal evaluation number being maxevals for each k point.

For a crystal, since this function does not know the volume of the Brillouin zone, it simply sums over k points and divide the result by the number of k points. The most common situation is calculating the density of ⟨O⟩, in which case the result should be multiplied by the volume of the Brillouin zone and divided by (2π)^3.

HopTB.Response.clt_Floquet_response — Function

clt_Floquet_response(getH::Function, getO::Function, getV::Function,
    getn::Function, l::Int64, ks::Matrix{Float64}; Γ::Float64=0.1,
    atol::Float64=0.0, maxevals::Int64=typemax(Int64))

Collect k-resolved Floquet response in ks. ks is in reduced coordinates and stored in columns.

Most options are identical to get_Floquet_response.