HopTB.Utilities

Introduction

This module contains some useful functions.

Unexported functions

HopTB.Utilities.integrate — Function

integrate(f::Function; rtol::Float64=√eps(),
    atol::Float64=0.0, maxevals::Int64=typemax(Int64))::Tuple

Integrate f from -∞ to ∞. The second argument of the returned tuple is error.

integrate(f::Function, x0::Vector{Float64}, x1::Vector{Float64};
    constant_components::Vector{Int64}=zeros(Int64, 0), rtol::Float64=√eps(),
    atol::Float64=0.0, maxevals::Int64=typemax(Int64), initdiv::Int64=1)

Integrate function f from x0 to x1 by HCubature package.

f is a function that takes a Vector{Float64} and yields a number or a matrix. The error of integartion is estimated by norm(..., Inf). atol and rtol are absolute tolerance and respective tolerance respectively. The integration will stop whenever atol or rtol is reached or when maxevals evaluations of f have been done. Initially, the volume is divided into initdiv segments along each dimension.

It is common that f is independent on some components of its argument x, in this case, one can specify these components by constant_components and the integration will generally be boosted.

HopTB.Utilities.pintegrate — Function

pintegrate(f::Function, x0::Vector{Float64}, x1::Vector{Float64};
    ndiv::Int64=1, constant_components::Vector{Int64}=zeros(Int64, 0),
    rtol::Float64=√eps(), atol::Float64=0.0, maxevals::Int64=typemax(Int64),
    initdiv::Int64=1)

Parallel integration of f. Most parameters have the same meaning as integrate. And the integration region is divided into ndiv segments in every dimension. The divided regions will be integrated on different processes.