HopTB.Utilities

Introduction

This module constructs Wannier functions.

exported functions

HopTB.Wannier.getwf — Function

getwf(getvs::Function, Rs::AbstractMatrix{Int64}; rtol::Float64=√eps(),
    atol::Float64=0.0, constant_components::Vector{Int64}=zeros(Int64, 0),
    ndiv::Int64=1)

Calculate Wannier functions in the home unit cell.

getvs should be a function that takes a three-component vector k as argument and yields a matrix (orbitalindex, bandindex). Wannier functions will only be computed at Rs. The other parameters are directly sent to HopTB.Wannier.pintegrate.

getwf(tm::TBModel, twfs::Dict{Vector{Int64}, Matrix{T}},
    bands::Vector{Int64}, Rs::AbstractMatrix{Int64}; tol::Float64=0.1, rtol::Float64=√eps(),
    atol::Float64=0.0, constant_components::Vector{Int64}=zeros(Int64, 0), ndiv::Int64=1) where T<:Number

Calculate Wannier functions with trial Wannier functions in twfs.

The project orthogonalization procedure has a tolerance tol. Other keyword arguments are directly sent to HopTB.Wannier.pintegrate.

getwf(getvs::Function, nkmesh::Vector{Int64}, nrmesh::Vector{Int64})

Calculate home unit cell Wannier functions with k mesh nkmesh.

Wannier functions are only computed at unit cells denoted by nrmesh.

getwf(tm::TBModel, twfs::Dict{Vector{Int64}, Matrix{T}},
    bands::Vector{Int64}, nkmesh::Vector{Int64},
    nrmesh::Vector{Int64}; tol::Float64=0.01) where T<:Number

Calculate Wannier functions for bands indexed by bands.

Both returned Wannier functions and lfs are stored in format {R: ⟨Rn|gm⟩}, where g are localized (Wannier) functions.

HopTB.Wannier.interpolatewf — Function

interpolatewf(gettm, tmparams::Matrix{<:Number},
    twfs::Dict{Vector{Int64}, Matrix{T}}, bands::Vector{Int64},
    nkmesh::Vector{Int64}, nrmesh::Vector{Int64}, ndiv::Int64) where T<:Number

Calculate Wannier functions by interpolation between tight binding models.

gettm is a function taking parameters and returning tight binding models. With the parameters stored in column in tmparams, the Wannier function is obtained by interpolation. More precisely, the twfs are used for initial guess and Wannier functions from last step is used as trial Wannier function for the next step.

Unexported functions

HopTB.Wannier.inner — Function

inner(vs::Matrix{ComplexF64}, lfs::Dict{Vector{Int64}, Matrix{T}},
    k::Vector{<:Real}) where T<:Number

Calculate ⟨ψnk|gm⟩, where g is a localized wave function. lfs is stored in format {R: ⟨Rn|gm⟩}.

HopTB.Wannier.project_orthogonalize — Function

project_orthogonalize(vs::Matrix{ComplexF64},
    lfs::Dict{Vector{Int64}, Matrix{T}}, k::Vector{<:Real};
    tol::Float64=0.01) where T<:Number

Get a new Bloch basis by project the Bloch wave functions onto localized functions.

The returned matrix stores ⟨ψm^(W)|φn⟩ in column, where φ_n is the new Bloch basis.